+++ src/utils/types.ts
@@ -1,4 +1,6 @@
 import { inferQueryOutput } from "./trpc";
+import { Role } from "@prisma/client";
+import { Status } from "@prisma/client";
 
 export type ProfileFormInputs = {
   firstName: string;
@@ -8,6 +10,30 @@ export type ProfileFormInputs = {
   companyName: string;
   companyAddress: string;
   status: "active" | "inactive";
+};
+
+// descripes a user's public data along with their POIs
+export type PublicUser = {
+  id: string;
+  name: string | null;
+  email: string | null;
+  image: string | null;
+  bio: string;
+  preferredName: string;
+  pronouns: string;
+  role: Role;
+  status: Status;
+  seatAvail: number;
+  companyName: string;
+  startPOILocation: string;
+  startPOICoordLng: number;
+  startPOICoordLat: number;
+  companyPOIAddress: string;
+  companyPOICoordLng: number;
+  companyPOICoordLat: number;
+  daysWorking: string;
+  startTime: Date | null;
+  endTime: Date | null;
 };
 
 export type User = inferQueryOutput<"user.me">;
